k8s-tests-project/pyproject.toml
[project]
name = "k8s-tests"
version = "0.1.0"
description = "Kubernetes E2E tests for Orange DevOps Task"
requires-python = ">=3.12"
dependencies = [
    "pytest>=7.4.0",
    "kubernetes>=28.1.0",
    "pyyaml>=6.0.1",
]
